# docs/swagger.yaml
basePath: /
definitions:
...
  dto.AttachmentResponse:
    properties:
      content_type:
        example: application/pdf
        type: string
      created_at:
        type: string
      file_name:
        example: syllabus.pdf
        type: string
      id:
        example: 1
        type: integer
      lesson_id:
        example: 1
        type: integer
      size:
        example: 20480
        type: integer
      updated_at:
        type: string
    type: object
  dto.ChapterResponse:
    properties:
      course_id:
        example: 1
        type: integer
      created_at:
        type: string
      description:
        example: Basic types, declaration, scope.
        type: string
      id:
        example: 1
        type: integer
      lessons:
        items:
          $ref: '#/definitions/dto.LessonResponse'
        type: array
      name:
        example: Variables and Types
        type: string
      order:
        example: 1
        type: integer
      updated_at:
        type: string
    type: object
...
  dto.CreateChapterRequest:
    properties:
      course_id:
        example: 1
        type: integer
      description:
        example: Basic types, declaration, scope.
        maxLength: 10000
        type: string
      name:
        example: Variables and Types
        maxLength: 255
        minLength: 2
        type: string
      order:
        example: 1
        minimum: 0
        type: integer
    required:
    - course_id
    - name
    type: object
...
  dto.ErrorBody:
    properties:
      code:
        example: COURSE_NOT_FOUND
        type: string
      message:
        example: course not found
        type: string
    type: object
  dto.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/dto.ErrorBody'
    type: object
  dto.LessonResponse:
    properties:
      chapter_id:
        example: 1
        type: integer
      content:
        example: In Go, var declares one or more variables...
        type: string
      created_at:
        type: string
      description:
        example: How to declare variables and infer types.
        type: string
      id:
        example: 1
        type: integer
      name:
        example: Declaring variables with var
        type: string
      order:
        example: 1
        type: integer
      updated_at:
        type: string
    type: object
  dto.UpdateChapterRequest:
    properties:
      description:
        example: Basic types, declaration, scope, examples.
        maxLength: 10000
        type: string
      name:
        example: Variables and Types (revised)
        maxLength: 255
        minLength: 2
        type: string
      order:
        example: 1
        minimum: 0
        type: integer
    required:
    - name
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: |-
    REST API for managing Courses, Chapters and Lessons in the LMS platform.
    All mutating endpoints (POST/PUT/DELETE) emit structured logs via logrus.
    Errors are returned in a unified shape: {"error":{"code":"...","message":"..."}}.
  title: LMS Main Service API
  version: "1.0"
paths:
...
  /chapters:
    post:
      consumes:
      - application/json
      description: Creates a chapter under an existing course. course_id is required
        and validated. Requires ROLE_ADMIN.
      parameters:
      - description: Chapter payload
        in: body
        name: chapter
        required: true
        schema:
          $ref: '#/definitions/dto.CreateChapterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.ChapterResponse'
        "400":
          description: Validation failed
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Missing or invalid token
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Requires ROLE_ADMIN
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Parent course not found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a new chapter (admin only)
      tags:
      - chapters
  /chapters/{id}:
    delete:
      description: Deletes a chapter. All lessons are cascade-deleted. Requires ROLE_ADMIN.
      parameters:
      - description: Chapter ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: Deleted
        "400":
          description: Invalid id
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Missing or invalid token
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Requires ROLE_ADMIN
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Chapter not found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete a chapter (admin only)
      tags:
      - chapters
    get:
      description: Returns a chapter together with its lessons, ordered by lesson.order.
      parameters:
      - description: Chapter ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ChapterResponse'
        "400":
          description: Invalid id
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Chapter not found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Get chapter by ID (with lessons)
      tags:
      - chapters
    put:
      consumes:
      - application/json
      description: Updates chapter name, description and order. course_id cannot be
        changed via PUT. Requires ROLE_ADMIN.
      parameters:
      - description: Chapter ID
        in: path
        name: id
        required: true
        type: integer
      - description: Chapter payload
        in: body
        name: chapter
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateChapterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ChapterResponse'
        "400":
          description: Validation failed or invalid id
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Missing or invalid token
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Requires ROLE_ADMIN
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Chapter not found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update an existing chapter (admin only)
      tags:
      - chapters
...
  /courses/{id}/chapters:
    get:
      description: Returns all chapters of the given course, ordered by chapter.order.
      parameters:
      - description: Course ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.ChapterResponse'
            type: array
        "400":
          description: Invalid id
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Course not found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: List chapters of a course
      tags:
      - chapters
  /download/{id}:
    get:
      description: Streams the stored file by attachment ID. Requires a valid token
        (any authenticated user).
      parameters:
      - description: Attachment ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/octet-stream
      responses:
        "200":
          description: OK
          schema:
            type: file
        "400":
          description: Invalid id
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Missing or invalid token
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Attachment not found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Download a file attachment
      tags:
      - attachments
...
  /lessons/{id}/attachments:
    get:
      description: Returns metadata for all files attached to the given lesson.
      parameters:
      - description: Lesson ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.AttachmentResponse'
            type: array
        "400":
          description: Invalid id
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Lesson not found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: List attachments of a lesson
      tags:
      - attachments
  /upload:
    post:
      consumes:
      - multipart/form-data
      description: 'Stores a file in MinIO and links it to a lesson. Multipart form:
        field "file" is the file, "lesson_id" the target lesson. Requires ROLE_ADMIN.'
      parameters:
      - description: Target lesson ID
        in: formData
        name: lesson_id
        required: true
        type: integer
      - description: File to upload
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.AttachmentResponse'
        "400":
          description: Missing file or invalid lesson_id
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Missing or invalid token
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Requires ROLE_ADMIN
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Lesson not found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Upload a file attachment (admin only)
      tags:
      - attachments
schemes:
- http
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and a JWT access token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"